use GPU timers instead of host timers

+++ src/framework/renderers/CLAbstractRenderer.h
@@ -212,7 +212,7 @@ protected:
     /**
      * @brief calculateRenderingTime_
      */
-    virtual double calculateRenderingTime_( );
+    virtual float calculateRenderingTime_( );
 
     /**
      * @brief defaultRenderingMode_
